+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Use codecs for FishIndex
-- Fix fishing_index not being found
+- Add compat with Aquaculture 2